backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
requests==2.31.0
pydantic==2.5.0
python-dateutil==2.8.2
APScheduler==3.10.4
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.5  # Für Zeitreihenvorhersagen
python-dotenv==1.0.0